move config to neptune package
