+++ server/neptune/config.go
@@ -1,4 +1,4 @@
-package config
+package neptune
 
 import (
 	"io"
