+++ server/neptune/temporalworker/server.go
@@ -20,7 +20,7 @@ import (
 	"github.com/runatlantis/atlantis/server/controllers"
 	"github.com/runatlantis/atlantis/server/controllers/templates"
 	"github.com/runatlantis/atlantis/server/logging"
-	"github.com/runatlantis/atlantis/server/neptune/config"
+	"github.com/runatlantis/atlantis/server/neptune"
 	neptune_http "github.com/runatlantis/atlantis/server/neptune/http"
 	"github.com/runatlantis/atlantis/server/neptune/temporal"
 	"github.com/runatlantis/atlantis/server/neptune/workflows"
@@ -49,7 +49,7 @@ type Server struct {
 	GithubActivities    *workflows.GithubActivities
 }
 
-func NewServer(config *config.Config) (*Server, error) {
+func NewServer(config *neptune.Config) (*Server, error) {
 	jobsController := &controllers.JobsController{
 		AtlantisVersion:     config.ServerCfg.Version,
 		AtlantisURL:         config.ServerCfg.URL,
